implement jwt token for fetching

+++ server/models/taskModel.js
@@ -1,17 +1,23 @@
 import pool from "../config/db.js";
 
-export const createTask = async (title, description, deadline, folderId) => {
+export const createTask = async (
+  title,
+  description,
+  deadline,
+  folderId,
+  userId
+) => {
   const [result] = await pool.query(
-    "INSERT INTO tasks (title, description, deadline, folder_id) VALUES (?, ?, ?, ?)",
-    [title, description, deadline, folderId]
+    "INSERT INTO tasks (title, description, deadline, folder_id, user_id) VALUES (?, ?, ?, ?, ?)",
+    [title, description, deadline, folderId, userId]
   );
   return result.insertId;
 };
 
-export const getTasksByFolder = async (folderId) => {
+export const getTasksByFolder = async (folderId, userId) => {
   const [rows] = await pool.query(
-    "SELECT * FROM tasks WHERE folder_id = ? ORDER BY position ASC",
-    [folderId]
+    "SELECT * FROM tasks WHERE folder_id = ? AND user_id = ? ORDER BY position ASC",
+    [folderId, userId]
   );
   return rows;
 };
@@ -48,22 +54,33 @@ export const deleteTask = async (taskId) => {
   await pool.query("DELETE FROM tasks WHERE id = ?", [taskId]);
 };
 
-export const createFolder = async (folderName) => {
-  const [result] = await pool.query("INSERT INTO folders (name) VALUES (?)", [
-    folderName,
-  ]);
+export const createFolder = async (folderName, user_id) => {
+  const [result] = await pool.query(
+    "INSERT INTO folders (name, user_id) VALUES (?, ?)",
+    [folderName, user_id]
+  );
   return result.insertId;
 };
 
-export const getFolders = async () => {
+export const getFolders = async (user_id) => {
   const [rows] = await pool.query(
-    "SELECT * FROM folders ORDER BY created_at ASC"
+    "SELECT * FROM folders WHERE user_id = ? ORDER BY created_at ASC",
+    [user_id]
   );
   return rows;
 };
 
-export const deleteFolder = async (folderId) => {
+export const deleteFolder = async (folderId, user_id) => {
+  const [folder] = await pool.query(
+    "SELECT id FROM folders WHERE id = ? AND user_id = ?",
+    [folderId, user_id]
+  );
+
+  if (folder.length === 0) return false;
+
   await pool.query("DELETE FROM tasks WHERE folder_id = ?", [folderId]);
 
   await pool.query("DELETE FROM folders WHERE id = ?", [folderId]);
+
+  return true;
 };
